+++ Products/CMFPlone/resources/browser/mixins.py
@@ -56,6 +56,7 @@ class LessConfiguration(BrowserView):
         result += "isPlone: true,\n"
         result += "isMockup: false,\n"
         result += "staticPath: '\"%s/++plone++static\"',\n" % site_url
+        result += "barcelonetaPath: '\"%s/++theme++barceloneta\"',\n" % site_url
 
         less_vars_params = {
             'site_url': site_url,
@@ -103,6 +104,7 @@ class LessModifyConfiguration(LessConfiguration):
         result2 += "'@isPlone': true,\n"
         result2 += "'@isMockup': false,\n"
         result2 += "'@staticPath: '\"%s/++plone++static\"',\n" % site_url
+        result2 += "'@barcelonetaPath: '\"%s/++theme++barceloneta\"',\n" % site_url
 
         less_vars_params = {
             'site_url': site_url,